Refactor to move activity
Add packages for activites and fragments.
