+++ app/src/androidTest/java/com/banderkat/recipes/RecipesMainActivityTest.java
@@ -5,6 +5,8 @@ import android.support.test.filters.LargeTest;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 
+import com.banderkat.recipes.activities.RecipesMainActivity;
+
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
