+++ app/src/main/res/layout/activity_recipes.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/recipe_activity_layout"
-    tools:context=".RecipesMainActivity">
+    tools:context=".activities.RecipesMainActivity">
 
     <TextView
         android:id="@+id/recipe_list_no_data"
